+++ example/src/pages/DropdownDemo.vue
@@ -1,0 +1,14 @@
+<template>
+  <Dropdown label="请点击">
+    <template #dropdown>
+      <DropdownItem>熊熊勇闯异世界</DropdownItem>
+      <DropdownItem>熊熊勇闯异世界</DropdownItem>
+      <DropdownItem>熊熊勇闯异世界</DropdownItem>
+      <DropdownItem>熊熊勇闯异世界</DropdownItem>
+    </template>
+  </Dropdown>
+</template>
+
+<script setup>
+import { Dropdown, DropdownItem } from '@apathia/apathia'
+</script>