+++ packages/cascader/demo/Basic.vue
@@ -1,0 +1,10 @@
+<template>
+  <ClientOnly>
+    <Cascader />
+  </ClientOnly>
+</template>
+
+<script setup>
+
+</script>
+