+++ packages/dropdown/src/types.ts
@@ -1,0 +1,5 @@
+import { ComputedRef } from 'vue'
+
+export interface DropdownInstance {
+  hideOnClick?: ComputedRef<boolean>
+}