+++ packages/popper/src/types.ts
@@ -1,3 +1,5 @@
+import { Instance } from '@popperjs/core'
+
 export type TriggerType = 'click' | 'hover' | 'focus' | 'manual'
 
 export const placement = [
@@ -16,3 +18,5 @@ export const placement = [
 ] as const
 
 export type Placement = typeof placement[number]
+
+export type PopperInstance = Instance