+++ packages/tag/src/types.ts
@@ -1,0 +1,10 @@
+export type TagType = 'success' | 'primary' | 'danger' | 'warning'
+
+export type TagItem =
+  | string
+  | {
+      type?: TagType
+      closeable?: boolean
+      iconClass?: string
+      [k: string]: any
+    }